Enable bugprone-misplaced-widening-cast
This clang-tidy check looks for operations where an arithmetic operation
is cast to a larger result.  E.g. static_cast<long>( a + b ) when a and
b are ints.  The expectation is that such casts are misplaced because if
a + b overflows, you would rather have cast it to a long before adding
b.

We had a few hits for this warning.  I don't think any of them were real
issues; mostly just casts for the sake of avoiding signed/unsigned
comparison warnings.  But I refactored them anyway.

+++ src/basecamp.cpp
@@ -145,9 +145,9 @@ expansion_data basecamp::parse_expansion( const std::string &terrain,
         const tripoint_abs_omt &new_pos )
 {
     expansion_data e;
-    int last_bar = terrain.find_last_of( '_' );
-    e.type = terrain.substr( base_camps::prefix_len, size_t( last_bar - base_camps::prefix_len ) );
-    e.cur_level = std::stoi( terrain.substr( size_t( last_bar + 1 ) ) );
+    size_t last_bar = terrain.find_last_of( '_' );
+    e.type = terrain.substr( base_camps::prefix_len, last_bar - base_camps::prefix_len );
+    e.cur_level = std::stoi( terrain.substr( last_bar + 1 ) );
     e.pos = new_pos;
     return e;
 }
